+++ src/components/ProcessingResults.vue
@@ -64,18 +64,7 @@
             </v-expansion-panel-text>
           </v-expansion-panel>
           <v-expansion-panel title="Processes" class="panel processes">
-            <v-expansion-panel-text>
-              <v-list density="compact" color="transparent" class="pa-0">
-                <v-list-item
-                  v-for="project in projects" 
-                  class="group" 
-                  :key="project"
-                  @click.stop="queryProcess(project)"
-                  >
-                    <v-label class="text-capitalize mb-1">{{ project }}</v-label>
-                </v-list-item>
-              </v-list>
-            </v-expansion-panel-text>
+            <ProcessList />
           </v-expansion-panel>
         </v-expansion-panels>
       </div>
@@ -111,11 +100,10 @@ import { ref, onMounted, onUnmounted, onBeforeUnmount, computed } from 'vue'
 import useMap from '../composables/useMap'
 import useNotifier from '../composables/useNotifier'
 import useAreaOfInterest from '../composables/useAreaOfInterest'
-import useProcessing from '../composables/useProcessing'
 import PropertiesDisplay from './PropertiesDisplay.vue'
+import ProcessList from './ProcessList.vue'
 import { mdiChevronDown, mdiTarget } from '@mdi/js'
 import { type Feature } from 'geojson'
-import { generateJWT } from '../functions/generate-jwt'
 
 const props = defineProps<{
   geoJsonResults: any[]
@@ -128,7 +116,6 @@ const emit = defineEmits<{
 const { map, handleMapClick, vectorLayer, selectedFeature, hidePropertiesBox } = useMap()
 const { clearResults, returnToResults, fitToExtent } = useAreaOfInterest()
 const { showInfo, showError } = useNotifier()
-const { projects } = useProcessing()
 
 const isOpen = ref(true)
 const limit = ref(50)
@@ -254,17 +241,6 @@ const downloadResults = () => {
     showError('Failed to download results.')
   }
 }
-
-const queryProcess = (id) => {
-  const token = generateJWT()
-  fetch(`${import.meta.env.VITE_API_BASE_URL}projects/${id}`, {
-    headers: {
-      'Content-Type': 'application/json',
-      Authorization: `Bearer ${token}`,
-    }
-  })
-}
-
 
 const fitMapToResult = (result: Feature) => {
   if (!result || !result.id) {
